+++ wallet/api/node/node.go
@@ -5,6 +5,7 @@ import (
 
 	apipb "code.vegaprotocol.io/vega/protos/vega/api/v1"
 	commandspb "code.vegaprotocol.io/vega/protos/vega/commands/v1"
+	"code.vegaprotocol.io/vega/wallet/api/node/adapters"
 )
 
 // Generates mocks
@@ -15,9 +16,8 @@ type Node interface {
 	Host() string
 	Stop() error
 	SendTransaction(context.Context, *commandspb.Transaction, apipb.SubmitTransactionRequest_Type) (string, error)
-	CheckTransaction(context.Context, *commandspb.Transaction) (*apipb.CheckTransactionResponse, error)
-	HealthCheck(context.Context) error
-	LastBlock(context.Context) (*apipb.LastBlockHeightResponse, error)
+	Statistics(ctx context.Context) (adapters.Statistics, error)
+	LastBlock(context.Context) (adapters.LastBlock, error)
 }
 
 // ReportType defines the type of event that occurred.